requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=0.19.0
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
selenium>=4.1.0
beautifulsoup4>=4.9.3
requests>=2.26.0
pydantic>=2.0.0